# Interfaces to abstract classes