#26 

Nets ~5% speedup on average for pending promises, just from changing interfaces to abstract classes.

Current master:

|                  Type |   N | BaseIsPending |       Mean | Code Size |
|---------------------- |---- |-------------- |-----------:|----------:|
|          AwaitPending | 100 |             ? | 424.685 μs |  16,647 B |
|          AsyncPending | 100 |             ? | 107.352 μs |   7,553 B |
|         AsyncResolved | 100 |             ? |  18.702 μs |   2,215 B |
|         AwaitResolved | 100 |             ? |  13.674 μs |   2,845 B |
|   ContinueWithPending | 100 |         False | 487.998 μs |  18,336 B |
| ContinueWithFromValue | 100 |         False |  24.998 μs |   7,238 B |
|  ContinueWithResolved | 100 |         False |  32.014 μs |   9,193 B |
|   ContinueWithPending | 100 |          True | 484.569 μs |  14,874 B |
| ContinueWithFromValue | 100 |          True | 128.037 μs |   9,317 B |
|  ContinueWithResolved | 100 |          True | 137.503 μs |   8,947 B |

This PR:

|                  Type |   N | BaseIsPending |      Mean | Code Size |
|---------------------- |---- |-------------- |----------:|----------:|
|          AwaitPending | 100 |             ? | 402.87 μs |  16,498 B |
|          AsyncPending | 100 |             ? | 101.79 μs |   7,565 B |
|         AsyncResolved | 100 |             ? |  18.13 μs |   2,215 B |
|         AwaitResolved | 100 |             ? |  13.45 μs |   2,845 B |
|   ContinueWithPending | 100 |         False | 465.93 μs |  18,891 B |
| ContinueWithFromValue | 100 |         False |  24.88 μs |   8,180 B |
|  ContinueWithResolved | 100 |         False |  31.76 μs |  10,187 B |
|   ContinueWithPending | 100 |          True | 463.06 μs |  15,416 B |
| ContinueWithFromValue | 100 |          True | 119.73 μs |   9,908 B |
|  ContinueWithResolved | 100 |          True | 138.84 μs |  10,026 B |